[BE-159] Separate integration tests into different crates
