+++ pallets/integration-tests/src/tests/controller_tests.rs
@@ -75,4 +75,119 @@ mod tests {
 				assert_ok!(MinterestProtocol::disable_collateral(alice(), CurrencyId::ETH));
 			});
 	}
+
+	// Extrinsic `set_insurance_factor`, description of scenario #2:
+	// Pool insurance is increased if the insurance_factor is greater than zero.
+	// 1. Alice deposit 40 DOT;
+	// 2. Alice borrow 20 DOT;
+	// 3. Set insurance factor equal 0.5.
+	// 4. Alice repay full loan in DOTs, pool insurance increased.
+	#[test]
+	fn set_insurance_factor_greater_than_zero() {
+		ExtBuilder::default()
+			.user_balance(ALICE, CurrencyId::DOT, ONE_HUNDRED)
+			.pool_user_data(CurrencyId::DOT, ALICE, BALANCE_ZERO, RATE_ZERO, true)
+			.pool_total_insurance(CurrencyId::DOT, BALANCE_ZERO)
+			.build()
+			.execute_with(|| {
+				// Alice deposit to DOT pool
+				let alice_deposited_amount = 40_000 * DOLLARS;
+				assert_ok!(MinterestProtocol::deposit_underlying(
+					Origin::signed(ALICE),
+					CurrencyId::DOT,
+					alice_deposited_amount
+				));
+
+				System::set_block_number(2);
+
+				// Alice borrow from DOT pool
+				let alice_borrowed_amount_in_dot = 20_000 * DOLLARS;
+				assert_ok!(MinterestProtocol::borrow(
+					Origin::signed(ALICE),
+					CurrencyId::DOT,
+					alice_borrowed_amount_in_dot
+				));
+
+				// Checking pool available liquidity
+				assert_eq!(
+					TestPools::get_pool_available_liquidity(CurrencyId::DOT),
+					alice_deposited_amount - alice_borrowed_amount_in_dot
+				);
+				// Checking total insurance for DOT pool.
+				assert_eq!(TestPools::pools(CurrencyId::DOT).total_insurance, BALANCE_ZERO);
+
+				System::set_block_number(10);
+
+				// Set insurance factor equal 0.5.
+				assert_ok!(TestController::set_insurance_factor(admin(), CurrencyId::DOT, 1, 2));
+
+				// Alice repay full loan in DOTs.
+				assert_ok!(MinterestProtocol::repay_all(Origin::signed(ALICE), CurrencyId::DOT));
+
+				let expected_interest_accumulated: Balance = 720_000_000_000_000;
+
+				// Checking pool available liquidity
+				assert_eq!(
+					TestPools::get_pool_available_liquidity(CurrencyId::DOT),
+					alice_deposited_amount + expected_interest_accumulated
+				);
+				assert_eq!(
+					TestPools::pools(CurrencyId::DOT).total_insurance,
+					BALANCE_ZERO + (expected_interest_accumulated / 2)
+				);
+			});
+	}
+
+	// Extrinsic `set_insurance_factor`, description of scenario #1:
+	// Pool insurance does not increase if the insurance_factor is zero.
+	// 1. Alice deposit 40 DOT;
+	// 2. Alice borrow 20 DOT;
+	// 3. Set insurance factor equal to zero.
+	// 4. Alice repay full loan in DOTs, pool total_insurance = 0.
+	#[test]
+	fn set_insurance_factor_equal_zero() {
+		ExtBuilder::default()
+			.user_balance(ALICE, CurrencyId::DOT, ONE_HUNDRED)
+			.pool_user_data(CurrencyId::DOT, ALICE, BALANCE_ZERO, RATE_ZERO, true)
+			.pool_total_insurance(CurrencyId::DOT, BALANCE_ZERO)
+			.build()
+			.execute_with(|| {
+				// Alice deposit to DOT pool
+				let alice_deposited_amount = 40_000 * DOLLARS;
+				assert_ok!(MinterestProtocol::deposit_underlying(
+					Origin::signed(ALICE),
+					CurrencyId::DOT,
+					alice_deposited_amount
+				));
+
+				System::set_block_number(2);
+
+				// Alice borrow from DOT pool
+				let alice_borrowed_amount_in_dot = 20_000 * DOLLARS;
+				assert_ok!(MinterestProtocol::borrow(
+					Origin::signed(ALICE),
+					CurrencyId::DOT,
+					alice_borrowed_amount_in_dot
+				));
+
+				// Checking pool available liquidity
+				assert_eq!(
+					TestPools::get_pool_available_liquidity(CurrencyId::DOT),
+					alice_deposited_amount - alice_borrowed_amount_in_dot
+				);
+				// Checking total insurance for DOT pool.
+				assert_eq!(TestPools::pools(CurrencyId::DOT).total_insurance, BALANCE_ZERO);
+
+				System::set_block_number(10);
+
+				// Set insurance factor equal to zero.
+				assert_ok!(TestController::set_insurance_factor(admin(), CurrencyId::DOT, 0, 1));
+
+				// Alice repay full loan in DOTs.
+				assert_ok!(MinterestProtocol::repay_all(Origin::signed(ALICE), CurrencyId::DOT));
+
+				// Checking pool total insurance.
+				assert_eq!(TestPools::pools(CurrencyId::DOT).total_insurance, BALANCE_ZERO);
+			});
+	}
 }
